-p (or \p) can now listen on a port range
-p cmd line option (or \p system command) can now listen on a port within a specified range e.g.
 q)\p 80/85
 q)\p
 81
range of ports is inclusive & tried in a random order. A service name can be used instead of a port number. The existing
option of using 0W to choose a free ephemeral port can be more efficient (where suitable).
range can be used in place of port number, when setting using existing rules e.g. for hostname
 q)\p myhost:2000/2010
or for multithreaded port
 q)\p -2000/2010