fixed broken links
Prev reported this on build
WARNING  -  Documentation file 'cloud/azure/architecture.md' contains a link to 'interfaces/fusion.md' which is not found in the documentation files.
WARNING  -  Documentation file 'ref/select.md' contains a link to 'q/basics/syscmds.md' which is not found in the documentation files.
WARNING  -  Documentation file 'ref/update.md' contains a link to 'q/basics/syscmds.md' which is not found in the documentation files.